refactor(compaction): drop the dead can_drop policy and unify part typing
CompactPolicyCanDrop was assigned but never consumed — droppability is
decided by the preserve-recent boundary and must-keep checks, so the flag
was a second source of truth waiting to drift. Tool part-type
classification also existed in three copies (candidate_rows, selection,
entry); consolidate on shared predicates next to entryPart. Split the
service machinery test file to stay within the file-size convention.
